Clean up and make export work
