--- src/containers/base/canvas/Container.ts
@@ -1,121 +1,0 @@
-import { getScreenDimension } from 'helpers';
-
-import { Arena, IArena } from 'containers/Arena/models/Arena';
-import { CellType, ICell } from 'containers/Arena/models/Cell';
-import {
-    BRICK_IMAGE,
-    EMPTY_BLACK_IMAGE,
-    EMPTY_IMAGE,
-    GRASS_IMAGE,
-    STEEL_IMAGE,
-    WATER_IMAGE,
-} from 'global/constants';
-
-export class CanvasBase {
-    private canvas: HTMLCanvasElement;
-    private context: CanvasRenderingContext2D | null;
-    private arena: IArena;
-    private imageMap: Map<string, HTMLImageElement> = new Map();
-
-    constructor() {
-        this.canvas = document.createElement('canvas');
-        this.context = this.canvas.getContext('2d');
-
-        this.setSize();
-        this.clearScene();
-        this.arena = new Arena();
-    }
-
-    public getCanvas = () => this.canvas;
-    public getContext = () => this.context;
-    public getArena = () => this.arena;
-    public getImageMap = () => this.imageMap;
-
-    protected getGameData = (): ICell[][] => this.arena.exportArenaData();
-    protected setGameData = (matrix: ICell[][]): void => {
-        this.arena.importArenaData(matrix);
-        this.renderScene();
-    }
-
-    protected renderScene = () => {
-        for (let i = 0; i < this.arena.size; i++) {
-            for (let j = 0; j < this.arena.size; j++) {
-                const cell = this.arena.matrix[i][j];
-                this.context!.drawImage(
-                    this.imageMap.get(cell.type)!,
-                    cell.position.x,
-                    cell.position.y,
-                    cell.size,
-                    cell.size,
-                );
-            }
-        }
-    }
-
-    protected getCellImage = (cellType: CellType): Promise<HTMLImageElement> => new Promise((resolve, reject) => {
-        const image = new Image();
-        switch (cellType) {
-            case CellType.BRICK:
-                image.src = BRICK_IMAGE;
-                break;
-            case CellType.GRASS:
-                image.src = GRASS_IMAGE;
-                break;
-            case CellType.STEEL:
-                image.src = STEEL_IMAGE;
-                break;
-            case CellType.WATER:
-                image.src = WATER_IMAGE;
-                break;
-            case CellType.EAGLE:
-                image.src = GRASS_IMAGE;
-                break;
-            case CellType.EMPTY:
-                image.src = EMPTY_IMAGE;
-                break;
-            case CellType.EMPTY_BLACK:
-                image.src = EMPTY_BLACK_IMAGE;
-                break;
-        }
-
-        image.onload = () => resolve(image);
-        image.onerror = (err) => reject(err);
-    })
-
-    private setSize = () => {
-        const { width, height } = getScreenDimension();
-        const size = width > height ? height : width;
-        this.canvas.width = size;
-        this.canvas.height = size;
-    }
-
-    private clearScene = () => {
-        this.context!.clearRect(0, 0, this.canvas.width, this.canvas.height);
-    }
-
-    private downloadTextures = async (): Promise<void> => {
-        const imagePromises = [
-            this.getCellImage(CellType.BRICK),
-            this.getCellImage(CellType.GRASS),
-            this.getCellImage(CellType.STEEL),
-            this.getCellImage(CellType.WATER),
-            this.getCellImage(CellType.EAGLE),
-            this.getCellImage(CellType.EMPTY_BLACK),
-        ];
-        const [
-            brickImage,
-            grassImage,
-            steelImage,
-            waterImage,
-            eagleImage,
-            emptyImageBlack,
-        ] = await Promise.all(imagePromises);
-        this.getImageMap().set(CellType.BRICK, brickImage);
-        this.getImageMap().set(CellType.GRASS, grassImage);
-        this.getImageMap().set(CellType.STEEL, steelImage);
-        this.getImageMap().set(CellType.WATER, waterImage);
-        this.getImageMap().set(CellType.EAGLE, eagleImage);
-        this.getImageMap().set(CellType.EMPTY, emptyImageBlack);
-        this.getImageMap().set(CellType.EMPTY_BLACK, emptyImageBlack);
-    }
-}
